requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.1.0
tldextract>=5.1.1
pandas>=2.2.2
tenacity>=8.3.0